Remove Feedback section and commence cleanup
