+++ public/index.html
@@ -4,23 +4,27 @@
     <meta charset="utf-8" />
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
 
-    <meta name="viewport" content="width=device-width,
-      initial-scale=1" />
+    <meta
+      name="viewport"
+      content="width=device-width,
+      initial-scale=1"
+    />
     <meta name="theme-color" content="#000000" />
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link rel="stylesheet"
+    <link
+      rel="stylesheet"
       href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap"
-      />
-    <link rel="stylesheet"
+    />
+    <link
+      rel="stylesheet"
       href="https://fonts.googleapis.com/css?family=Rubik:300,400,500&display=swap"
-      />
+    />
 
     <title>IDYEAHS</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-
   </body>
 </html>
