+++ src/app/App.js
@@ -9,8 +9,6 @@ import {
 } from '../utils/utils';
 import GlobalStyles from '../misc/GlobalStyles';
 import mockIdeas from '../data/MockIdeasData';
-import feedback from '../data/Feedback';
-import responsesData from '../data/Responses';
 import usersList from '../data/Users';
 import Home from './Home';
 import IdeasFeed from './IdeasFeed';
@@ -21,8 +19,6 @@ import IdeaDetailsView from './IdeaDetailsView';
 import Login from './Login';
 import SignUp from './SignUp';
 import SignUpSuccess from './SignUpSuccess';
-import FeedbackResultsPage from './FeedbackResults';
-import FeedbackForm from './FeedbackForm';
 import UserPage from './UserProfile';
 
 const Grid = styled.div`
@@ -33,9 +29,6 @@ const Grid = styled.div`
 
 export default function App() {
   const [ideas, setIdeas] = useState(getLocal('ideas') || mockIdeas);
-  const [responses, setResponses] = useState(
-    getLocal('responses') || responsesData
-  );
   const [users, setUsers] = useState(getLocal('users') || usersList);
   const [isLoggedIn, setIsLoggedIn] = useState(getLocal('isLoggedIn') || false);
   const [activeUser, setActiveUser] = useState(getLocal('activeUser') || []);
@@ -44,7 +37,6 @@ export default function App() {
   useEffect(() => setLocal('users', users), [users]);
   useEffect(() => setLocal('isLoggedIn', isLoggedIn), [isLoggedIn]);
   useEffect(() => setLocal('ideas', ideas), [ideas]);
-  useEffect(() => setLocal('responses', responses), [responses]);
 
   function handleIdeaSubmit(newIdeas, history) {
     setIdeas(newIdeas);
@@ -111,14 +103,6 @@ export default function App() {
     ];
     newIdeas[ideasIndex].stars = newStars;
     setIdeas(newIdeas);
-  }
-
-  function handleFeedbackSubmit(response, history) {
-    // const responsesSum = responses.map(
-    //   (number, index) => Number(number) + Number(answers[index])
-    // );
-    setResponses([...responses, response]);
-    history.push('/feedback');
   }
 
   function handleLogin(loggedInUser, history) {
@@ -242,31 +226,6 @@ export default function App() {
               onStarAdd={handleStarAdd}
               onStarRemove={handleStarRemove}
               activeUser={activeUser}
-            />
-          )}
-        />
-
-        <Route
-          exact
-          path="/feedback"
-          render={props => (
-            <FeedbackResultsPage
-              heading="Feedback"
-              responses={responses}
-              questions={feedback}
-              activeUser={activeUser}
-              history={props.history}
-            />
-          )}
-        />
-        <Route
-          exact
-          path="/feedback/add"
-          render={props => (
-            <FeedbackForm
-              questions={feedback}
-              handleFeedbackSubmit={handleFeedbackSubmit}
-              history={props.history}
             />
           )}
         />
