+++ src/app/Buttons.elements.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+import SubmitButton from '../components/form/SubmitButton'
+
+export const ButtonTransparent = styled(SubmitButton)`
+  background: transparent;
+  border: 1px solid #008dff;
+  color: #008dff;
+`;
+
+export const ButtonFilled = styled(SubmitButton)`
+  background: white;
+  color: #008dff;
+  border: none;
+`;
