--- src/data/Feedback.js
@@ -1,84 +1,0 @@
-const feedback = [
-  {
-    id: 0,
-    question_text_en: 'Do you know what is expected of you at work?',
-    question_text_de:
-      'Mir ist klar, was von mir als Arbeitsleistung erwartet wird.',
-  },
-  {
-    id: 1,
-    question_text_en:
-      'Do you have the materials and equipment to do your work right?',
-    question_text_de:
-      'Ich verfüge über die nötigen Arbeitsmaterialien und Arbeitsbedingungen, um meine Arbeit gut und richtig auszuführen.',
-  },
-  {
-    id: 2,
-    question_text_en:
-      'At work, do you have the opportunity to do what you do best every day?',
-    question_text_de:
-      'Bei der Arbeit habe ich stets die Gelegenheit, das zu tun, was ich am besten kann.',
-  },
-  {
-    id: 3,
-    question_text_en:
-      'In the last seven days, have you received recognition or praise for doing good work?',
-    question_text_de:
-      'In den letzten sieben Tagen habe ich für gute Arbeit Anerkennung oder Lob erhalten.',
-  },
-  {
-    id: 4,
-    question_text_en:
-      'Does your supervisor, or someone at work, seem to care about you as a person?',
-    question_text_de:
-      'Meine Vorgesetzte/n oder jemand anders bei der Arbeit interessiert und schätzt mich als Mensch.',
-  },
-  {
-    id: 5,
-    question_text_en:
-      'Is there someone at work who encourages your development?',
-    question_text_de:
-      'Bei der Arbeit gibt es jemanden, der mich in meiner Entwicklung fördert.',
-  },
-  {
-    id: 6,
-    question_text_en: 'At work, do your opinions seem to count?',
-    question_text_de: 'In meinem Arbeitsumfeld hat meine Meinung Gewicht.',
-  },
-  {
-    id: 7,
-    question_text_en:
-      'Does the mission/purpose of your company make you feel your job is important?',
-    question_text_de:
-      'Das Ziel und die Unternehmensphilosophie unserer Firma geben mir das Gefühl, dass meine Arbeit wichtig ist.',
-  },
-  {
-    id: 8,
-    question_text_en:
-      'Are your associates (fellow employees) committed to doing quality work?',
-    question_text_de:
-      'Meine Kollegen / Kolleginnen fühlen sich verpflichtet und verantwortlich, Qualität in ihrer Arbeit abzuliefern.',
-  },
-  {
-    id: 9,
-    question_text_en: 'Do you have a best friend at work?',
-    question_text_de:
-      'In meiner Firma habe ich einen guten Freund / eine gute Freundin.',
-  },
-  {
-    id: 10,
-    question_text_en:
-      'In the last six months, has someone at work talked to you about your progress?',
-    question_text_de:
-      'Im vergangenen Halbjahr hat jemand in meinem Unternehmen mit mir über meine Fortschritte gesprochen.',
-  },
-  {
-    id: 11,
-    question_text_en:
-      'In the last year, have you had opportunities to learn and grow?',
-    question_text_de:
-      'Ich hatte im vergangenen Jahr die Gelegenheit, in meinem Unternehmen dazuzulernen und mich weiterzuentwickeln.',
-  },
-];
-
-export default feedback;
