--- src/data/Responses.js
@@ -1,17 +1,0 @@
-import uid from 'uid';
-
-const responsesData = [
-  {
-    id: uid(),
-    date: '2019-06-08T08:02:17-05:00',
-    results: [3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3],
-    counter: '1',
-  },
-  {
-    id: uid(),
-    date: '2019-06-01T08:02:17-05:00',
-    results: [3, 4, 5, 2, 5, 3, 6, 7, 8, 7, 5, 6],
-    counter: '2',
-  },
-];
-export default responsesData;
