--- src/wip/wip.css
@@ -1,3 +1,0 @@
-.header-shadow {
-  box-shadow: 0 -4px 22px 3px lightslategray;
-}
